fixed changing email problem
